Use custom colors in stats
Closes #211
